+++ docs/changelog.md
@@ -11,8 +11,9 @@
 - Add support for prometheus metrics.
 - Add support for API Key generation.
 - Extend ExVe model with Fleet APIs.
-- Add `ignitionStatuses`, `fuelLevels`, and `doorStatuses` endpoints to ExVe
-  API.
+- Add `ignitionStatuses`, `fuelLevels`, `doorStatuses`, `highBeamStatuses`,
+  `headlampStatuses`, `parkingBrakeStatuses`, and `windshieldWiperStatuses`
+  endpoints to ExVe API.
 
 ## 1.1.1